refactor(rpc): move chain handlers to routes/chain.rs (backlog #12 phase 2d)
Fourth routes slice. Four chain-wide endpoints out:

  - chain_info       (GET /chain/info)
  - get_blocks       (GET /chain/blocks, paginated)
  - get_block        (GET /chain/blocks/{index})
  - validate_chain   (GET /chain/validate, auth-gated)

validate_chain's result cache (VALIDATE_CACHE_HEIGHT +
VALIDATE_CACHE_RESULT atomics) moves with the handler, along with
the two M-07 cache tests that covered it.

Zero route / behaviour change. Remaining slices: accounts,
transactions, epoch.
